Attend to comments
