+++ integrations/github/.port/resources/blueprints.json
@@ -113,7 +113,6 @@
             }
         }
     },
-
     {
         "identifier": "githubUser",
         "title": "Github User",
